type fixes

+++ src/kskm/keymaster/delete.py
@@ -23,13 +23,13 @@ def key_delete(label: str, p11modules: KSKM_P11, force: bool=False) -> bool:
 
     logger.info(f'Deleting key pair {existing_key}')
     session = get_session(p11modules, logger)
-    if existing_key.public_key:
+    if existing_key.public_key and existing_key.pubkey_handle:
         res = session.destroyObject(existing_key.pubkey_handle[0])
         logger.debug(f'Public key C_DestroyObject result: {res}')
 
     # Handles seem to get invalidated when calling destroyObject, so do another search for a private key
     existing_key = get_p11_key(label, p11modules, public=False)
-    if existing_key.privkey_handle:
+    if existing_key and existing_key.privkey_handle and existing_key.privkey_handle:
         res = session.destroyObject(existing_key.privkey_handle[0])
         logger.debug(f'Private key C_DestroyObject result: {res}')
         return True
